non-error-backside/requirements.txt
pip freeze > requirements.txt
django
djangorestframework
gunicorn
mysqlclient    # or psycopg2 if using PostgreSQL
python-dotenv
